IO-300 FileUtils.moveDirectoryToDirectory removes source directory if destination is a subdirectory
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/io/trunk@1300447 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/io/FileUtils.java
@@ -32,6 +32,7 @@ import java.net.URLConnection;
 import java.nio.ByteBuffer;
 import java.nio.channels.FileChannel;
 import java.nio.charset.Charset;
+import java.security.InvalidParameterException;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
@@ -2491,6 +2492,9 @@ public class FileUtils {
         }
         boolean rename = srcDir.renameTo(destDir);
         if (!rename) {
+            if (destDir.getCanonicalPath().startsWith(srcDir.getCanonicalPath())) {
+                throw new IOException("Cannot move directory: "+srcDir+" to a subdirectory of itself: "+destDir);
+            }
             copyDirectory( srcDir, destDir );
             deleteDirectory( srcDir );
             if (srcDir.exists()) {
